requirements.txt
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.5.0
asttokens==2.4.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.24.75
botocore==1.27.75
Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.2
fasteners==0.14.1
fonttools==4.50.0
future==0.18.2
google-auth==2.28.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jeepney==0.7.1
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==12.0.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==1.3.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
systemd-python==234
threadpoolctl==3.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.8.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.23.2
uvloop==0.18.0
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
